+++ HISTORY.rst
@@ -415,6 +415,8 @@ Added
 - unittests in "tests/test_config.py", which are concerned with testing the functionality exposed by the config module
 - Added "FigureTestResult" which allows to add a matplotlib figure directly as the result of a test.
 - Added camera test "RepeatedResetTest"
+- Added camera test "RepeatedFrameTest"
+- Added camera test "CalculatePairNoiseTest"
 
 Changes